+++ integrations/github/github/webhook/webhook_processors/file_webhook_processor.py
@@ -20,6 +20,10 @@ from github.helpers.utils import ObjectKind
 from integration import GithubFilePattern, GithubFileResourceConfig
 from loguru import logger
 from github.helpers.port_app_config import ORG_CONFIG_REPO
+from github.enrichments.included_files import (
+    IncludedFilesEnricher,
+    FileIncludedFilesStrategy,
+)
 
 YAML_SUFFIX = (".yaml", ".yml")
 JSON_SUFFIX = ".json"
@@ -85,14 +89,15 @@ class FileWebhookProcessor(BaseRepositoryWebhookProcessor):
             current_branch,
         )
 
-        included_files = selector.included_files or []
-        if included_files and updated_raw_results:
-            from main import _enrich_file_entities_batch_with_included_files
-
+        if updated_raw_results and selector.included_files:
             rest_client = create_github_client()
-            updated_raw_results = await _enrich_file_entities_batch_with_included_files(
-                rest_client, updated_raw_results, included_files
+            enricher = IncludedFilesEnricher(
+                rest_client=rest_client,
+                strategy=FileIncludedFilesStrategy(
+                    included_files=selector.included_files,
+                ),
             )
+            updated_raw_results = await enricher.enrich_batch(updated_raw_results)
 
         return WebhookEventRawResults(
             updated_raw_results=updated_raw_results,